+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -142,6 +142,12 @@ Improvements to clang-tidy
   Detects local variable declarations declaring more than one variable and
   tries to refactor the code to one statement per declaration.
 
+- New :doc:`readability-const-return-type
+  <clang-tidy/checks/readability-const-return-type>` check.
+
+  Checks for functions with a ``const``-qualified return type and recommends
+  removal of the ``const`` keyword.
+
 - New :doc:`readability-magic-numbers
   <clang-tidy/checks/readability-magic-numbers>` check.